未测试！！
Enemy信息缩放参数
FrontDisplay缩放参数
Player蓄力开始常数外置，Player为IPhone的移动
为tablefile名增加MakePath
增加IPhone矩阵和IPhone用Super矩阵
resourcefilename为空判断增加单空格
增加InitDllLoader的__WIN32判断
Main增加IPhone回调，整理初始化、循环、结束函数
IPhone的touch处理，增加ini屏幕和信息缩放参数
增加IPhone的前端显示sprite
Const增加IPhone屏幕常数
增加IPhone的musicData结构对齐调整
调整clientAfterInitial并增加ResetMatrix
增加IPhone专用的SI常数字符串
增加IPhone初始化等class文件
更改获得ini文件直接使用常数的错误
增加IPhone用PlayerSelect一直按住的按键
在startPrep处存储lastmatchchara
增加IPhone的Render

修正exception中的WIN32判断

修正LuaPlus对IPhone的误判，修正其中的GetInteger方法
修改ldump中的无名union
增加loadlib中的显示强制转换

增加IPhone用资源

增加IPhone的Graphic处理
修正Header引用的斜杠
分割System_Start为三部分，增加HGE_MANAGELOOP设置
增加IPhone需要的strupr和strcmpi
增加IPhone的sound支持
增加IPhone时间处理函数

修改zlib中的DWORD定义
增加IPhone需要的库

修改PSP_framebuffer适合IPhone
修正PSP_Graphics可能隐含的重定义错误

+++ LuaPlus/LuaPlus/LuaObject.cpp
@@ -409,7 +409,7 @@ size_t LuaObject::ToStrLen()
 int LuaObject::GetInteger() const
 {
 	luaplus_assert(m_state  &&  IsInteger());
-	return (int)(unsigned int)nvalue(&m_object);
+	return (int)/*(unsigned int)*/nvalue(&m_object);
 }
